Start of Project, main html, css, and java script
Started project by creating the base html page to contain the forms that direct to the blog page.  Forms will be added next.

Created the styling for the sheet.  Simple body containers with a header, asides and footer.

Created light and dark mode toggle.

Finished css and light/dark mode first so that I can copy to the blog page and do less work.