Fixed Comments, removed old css and updated index with button
Spent some time fixing and organizing the comments a little better.
Removed some old css that was still around from the first iteration of the light and dark mode.  Added a button to the index page to go to the blog allowing one to go back and forth between the two easily.