Re-adjusted JavaScript files to match names to usage.  Added in local storage options and tests.
I re-adjusted the JavaScript files to match names to usage.  Form was being used for light dark mode instead of the local storage.  Added that to logic and created two new files form and form blog for the push and pull of storage objects.

Added in local storage options and tests. First iteration to debug any problems with converting school examples to working code within this blog project.