style.css for the index.html changes
Added css styling changes to adjust the blog entry form and fields appearance.
They were not located in the correct positions and had differing styles compared to the rest of the page.

Added media query to adjust the aside 1 and 2 flex control to stack when below 768px screen width to avoid squishing issues.